test(web): migrate Diff page to DOM-free spec + Web e2e
Delete the 810-line diff.page.spec.ts plus the diff-result-renderer,
diff-hunk, and source-snippet component specs (1913 lines of DOM
assertions in total) and the data-testid attributes that only fed
them. Add a Web e2e covering a graph diff (alpha.ttl vs beta.ttl ->
+5 -5 hunks) and a tabular diff (matching SELECT ?s projections ->
+2 -2 rows) against the existing fixture sources.

Extract computeTotalsLine() out of the renderer into a pure util with
a DOM-free spec; add a direct spec for formatDiffError /
formatTargetError. Mode dispatch and mixed-shape rejection remain
asserted server-side in libs/server/src/lib/diff/diff.service.spec.ts;
the renderer just dispatches on result.kind. Source-snippet windowing
and hunk classification were already covered by their pure utils'
existing specs.

Reorganize the diff/ folder to mirror describe/ and sources/ —
diff.service.ts now holds only the HTTP wrapper; the response and
error wire shapes live in models/, and the error formatters live in
utils/. Surviving data-testids on the Diff page (editor-left/right,
diff-totals, tabular-diff, hunk) are the ones the new e2e relies on.
